Added navbar to sign up/sign in page. Put profile card components into seperate files
