+++ immichat-frontend/src/components/ProfileComponents/FirstName.js
@@ -1,0 +1,16 @@
+const FirstName = ({ onChange, value }) => (
+  <div className="field">
+    <label htmlFor="name">First Name:</label>
+    <input
+      id="fname"
+      type="text"
+      onChange={onChange}
+      maxLength="25"
+      value={value}
+      placeholder="First Name"
+      required
+    />
+  </div>
+);
+
+export default FirstName;
